Move `register_controller` into `base_test`.
So the controller objects of a class is encapsulatd within the test
class itself. This makes the structure cleaner and gets rid of the
awkward passing of a partial function from test runner to each test
class.
