+++ tests/mobly/base_test_test.py
@@ -29,6 +29,7 @@ from mobly import records
 from mobly import signals
 
 from tests.lib import utils
+from tests.lib import mock_controller
 
 MSG_EXPECTED_EXCEPTION = "This is an expected exception."
 MSG_EXPECTED_TEST_FAILURE = "This is an expected test failure."
@@ -45,6 +46,15 @@ class SomeError(Exception):
     """A custom exception class used for tests in this module."""
 
 
+class MockEmptyBaseTest(base_test.BaseTestClass):
+    """Stub used to test functionalities not specific to a class
+    implementation.
+    """
+
+    def test_func(self):
+        pass
+
+
 class BaseTestTest(unittest.TestCase):
     def setUp(self):
         self.tmp_dir = tempfile.mkdtemp()
@@ -52,6 +62,7 @@ class BaseTestTest(unittest.TestCase):
         self.summary_file = os.path.join(self.tmp_dir, 'summary.yaml')
         self.mock_test_cls_configs.summary_writer = records.TestSummaryWriter(
             self.summary_file)
+        self.mock_test_cls_configs.controller_configs = {}
         self.mock_test_cls_configs.log_path = self.tmp_dir
         self.mock_test_cls_configs.user_params = {"some_param": "hahaha"}
         self.mock_test_cls_configs.reporter = mock.MagicMock()
@@ -1772,6 +1783,113 @@ class BaseTestTest(unittest.TestCase):
                 self.assertEqual(c['a'], content['a'])
                 self.assertIsNotNone(c['timestamp'])
         self.assertTrue(hit)
+
+    def test_register_controller_no_config(self):
+        bt_cls = MockEmptyBaseTest(self.mock_test_cls_configs)
+        with self.assertRaisesRegex(signals.ControllerError,
+                                    'No corresponding config found for'):
+            bt_cls.register_controller(mock_controller)
+
+    def test_register_controller_no_config_for_not_required(self):
+        bt_cls = MockEmptyBaseTest(self.mock_test_cls_configs)
+        self.assertIsNone(
+            bt_cls.register_controller(mock_controller, required=False))
+
+    def test_register_controller_dup_register(self):
+        """Verifies correctness of registration, internal tally of controllers
+        objects, and the right error happen when a controller module is
+        registered twice.
+        """
+        mock_test_config = self.mock_test_cls_configs.copy()
+        mock_ctrlr_config_name = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+        mock_test_config.controller_configs = {
+            mock_ctrlr_config_name: ['magic1', 'magic2']
+        }
+        bt_cls = MockEmptyBaseTest(mock_test_config)
+        bt_cls.register_controller(mock_controller)
+        registered_name = 'mock_controller'
+        self.assertTrue(registered_name in bt_cls._controller_registry)
+        mock_ctrlrs = bt_cls._controller_registry[registered_name]
+        self.assertEqual(mock_ctrlrs[0].magic, 'magic1')
+        self.assertEqual(mock_ctrlrs[1].magic, 'magic2')
+        self.assertTrue(bt_cls._controller_destructors[registered_name])
+        expected_msg = 'Controller module .* has already been registered.'
+        with self.assertRaisesRegex(signals.ControllerError, expected_msg):
+            bt_cls.register_controller(mock_controller)
+
+    def test_register_controller_no_get_info(self):
+        mock_test_config = self.mock_test_cls_configs.copy()
+        mock_ctrlr_config_name = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+        get_info = getattr(mock_controller, 'get_info')
+        delattr(mock_controller, 'get_info')
+        try:
+            mock_test_config.controller_configs = {
+                mock_ctrlr_config_name: ['magic1', 'magic2']
+            }
+            bt_cls = MockEmptyBaseTest(mock_test_config)
+            bt_cls.register_controller(mock_controller)
+            self.assertEqual(bt_cls.results.controller_info, {})
+        finally:
+            setattr(mock_controller, 'get_info', get_info)
+
+    def test_register_controller_return_value(self):
+        mock_test_config = self.mock_test_cls_configs.copy()
+        mock_ctrlr_config_name = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+        mock_test_config.controller_configs = {
+            mock_ctrlr_config_name: ['magic1', 'magic2']
+        }
+        bt_cls = MockEmptyBaseTest(mock_test_config)
+        magic_devices = bt_cls.register_controller(mock_controller)
+        self.assertEqual(magic_devices[0].magic, 'magic1')
+        self.assertEqual(magic_devices[1].magic, 'magic2')
+
+    def test_register_controller_change_return_value(self):
+        mock_test_config = self.mock_test_cls_configs.copy()
+        mock_ctrlr_config_name = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+        mock_test_config.controller_configs = {
+            mock_ctrlr_config_name: ['magic1', 'magic2']
+        }
+        bt_cls = MockEmptyBaseTest(mock_test_config)
+        magic_devices = bt_cls.register_controller(mock_controller)
+        magic1 = magic_devices.pop(0)
+        self.assertIs(magic1,
+                      bt_cls._controller_registry['mock_controller'][0])
+        self.assertEqual(
+            len(bt_cls._controller_registry['mock_controller']), 2)
+
+    def test_register_controller_less_than_min_number(self):
+        mock_test_config = self.mock_test_cls_configs.copy()
+        mock_ctrlr_config_name = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+        mock_test_config.controller_configs = {
+            mock_ctrlr_config_name: ['magic1', 'magic2']
+        }
+        bt_cls = MockEmptyBaseTest(mock_test_config)
+        expected_msg = 'Expected to get at least 3 controller objects, got 2.'
+        with self.assertRaisesRegex(signals.ControllerError, expected_msg):
+            bt_cls.register_controller(mock_controller, min_number=3)
+
+    def test_verify_controller_module(self):
+        base_test._verify_controller_module(mock_controller)
+
+    def test_verify_controller_module_null_attr(self):
+        try:
+            tmp = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+            mock_controller.MOBLY_CONTROLLER_CONFIG_NAME = None
+            msg = 'Controller interface .* in .* cannot be null.'
+            with self.assertRaisesRegex(signals.ControllerError, msg):
+                base_test._verify_controller_module(mock_controller)
+        finally:
+            mock_controller.MOBLY_CONTROLLER_CONFIG_NAME = tmp
+
+    def test_verify_controller_module_missing_attr(self):
+        try:
+            tmp = mock_controller.MOBLY_CONTROLLER_CONFIG_NAME
+            delattr(mock_controller, 'MOBLY_CONTROLLER_CONFIG_NAME')
+            msg = 'Module .* missing required controller module attribute'
+            with self.assertRaisesRegex(signals.ControllerError, msg):
+                base_test._verify_controller_module(mock_controller)
+        finally:
+            setattr(mock_controller, 'MOBLY_CONTROLLER_CONFIG_NAME', tmp)
 
 
 if __name__ == "__main__":
